+++ interfaces/doc.go
@@ -1,4 +1,0 @@
-// Package interfaces contains deprecated compatibility aliases.
-//
-// Deprecated: use domain packages such as activation and websearch.
-package interfaces