+++ interfaces/toolactivation.go
@@ -1,8 +1,0 @@
-package interfaces
-
-import "github.com/codewandler/agentsdk/activation"
-
-// ActivationState manages which tools are active/inactive.
-//
-// Deprecated: use activation.State.
-type ActivationState = activation.State